# src/main/resources/assets/swagger/api-docs/coralApi.yaml
basePath: /v0/
definitions:
  Account:
    properties:
      active:
        type: boolean
      bdate:
        format: date-time
        type: string
      description:
        type: string
      edate:
        format: date-time
        type: string
      name:
        type: string
      organization:
        type: string
      type:
        type: string
  AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
  DisableRequest:
    properties:
      item:
        type: string
  EnableRequest:
    properties:
      item:
        type: string
      member:
        type: string
      project:
        type: string
  LabRoles:
    properties:
      empty:
        type: boolean
  Machines:
    properties:
      empty:
        type: boolean
  Member:
    properties:
      active:
        type: boolean
      address1:
        type: string
      address2:
        type: string
      advisor:
        type: string
      altFax:
        type: string
      altOffice:
        type: string
      altPhone:
        type: string
      bdate:
        format: date-time
        type: string
      city:
        type: string
      disability:
        type: string
      edate:
        format: date-time
        type: string
      email:
        type: string
      ethnicity:
        type: string
      fax:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      mailCode:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      project:
        type: string
      race:
        type: string
      state:
        type: string
      type:
        type: string
      univid:
        type: string
      url:
        type: string
      zipcode:
        type: string
  PasswordResetRequest:
    properties:
      name:
        type: string
      password:
        type: string
  Project:
    properties:
      account:
        type: string
      active:
        type: boolean
      bdate:
        format: date-time
        type: string
      description:
        type: string
      discipline:
        type: string
      edate:
        format: date-time
        type: string
      name:
        type: string
      nickname:
        type: string
      'null':
        type: boolean
      pi:
        type: string
      type:
        type: string
  ProjectMembership:
    properties:
      members:
        items:
          type: string
        type: array
      project:
        type: string
  Projects:
    properties:
      empty:
        type: boolean
  ReservationRequest:
    properties:
      bdate:
        type: string
      item:
        type: string
      lengthInMinutes:
        format: int32
        type: integer
      member:
        type: string
      project:
        type: string
  User:
    properties:
      username:
        type: string
host: 'ryanmb.nanofab.utah.edu:4001'
info:
  title: CoralAPIServer
  version: '0.3.3'
paths:
  account:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: name
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Account'
      summary: Find account by name
      tags:
        - account
    post:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Account'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Account'
      summary: update
      tags:
        - account
    put:
      description: ''
      operationId: updateRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Account'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Account'
      summary: update
      tags:
        - account
  authenticate:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: proxyFor
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
      tags:
        - authenticate
    post:
      description: ''
      operationId: authRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/AuthRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
      tags:
        - authenticate
  checkKey:
    get:
      description: ''
      operationId: checkKey
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
      tags:
        - checkKey
  disable:
    post:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/DisableRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/DisableRequest'
      tags:
        - disable
  enable:
    post:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/EnableRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/EnableRequest'
      tags:
        - enable
  labRoles:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: member
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/LabRoles'
      tags:
        - labRoles
  machine:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: name
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Machines'
      tags:
        - machine
  member:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: name
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Member'
      summary: Find member by name
      tags:
        - member
    post:
      description: ''
      operationId: updateRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Member'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Member'
      summary: Update member
      tags:
        - member
    put:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Member'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Member'
      summary: Create member
      tags:
        - member
  project:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: name
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Project'
      summary: Find project by name
      tags:
        - project
    post:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Project'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Project'
      summary: Update project
      tags:
        - project
    put:
      description: ''
      operationId: updateRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Project'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Project'
      summary: Create project
      tags:
        - project
  projectMembership:
    get:
      description: ''
      operationId: get
      parameters:
        - in: query
          name: project
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/ProjectMembership'
      tags:
        - projectMembership
    put:
      description: ''
      operationId: update
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/ProjectMembership'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/ProjectMembership'
      tags:
        - projectMembership
  projects:
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: member
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/Projects'
      tags:
        - projects
  reservation:
    delete:
      description: ''
      operationId: deleteRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/ReservationRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/ReservationRequest'
      tags:
        - reservation
    get:
      description: ''
      operationId: getRequest
      parameters:
        - in: query
          name: machine
          required: false
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/ReservationRequest'
      tags:
        - reservation
    post:
      description: ''
      operationId: createRequest
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/ReservationRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/ReservationRequest'
      tags:
        - reservation
  resetPassword:
    post:
      description: ''
      operationId: updatePassword
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/PasswordResetRequest'
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
      tags:
        - resetPassword
  version:
    get:
      description: ''
      operationId: getRequest
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
      tags:
        - version
  whoami:
    get:
      description: ''
      operationId: getRequest
      produces:
        - application/json
      responses:
        '200':
          description: No response was specified
          schema:
            $ref: '#/definitions/User'
      tags:
        - whoami
schemes:
  - http
swagger: '2.0'
tags:
  - name: enable
  - name: authenticate
    description: ''
  - name: resetPassword
  - name: member
    description: ''
  - name: account
    description: ''
  - name: projects
  - name: whoami
  - name: version
  - name: reservation
  - name: checkKey
    description: check if key is working in coral
  - name: project
    description: ''
  - name: projectMembership
  - name: labRoles
  - name: disable
  - name: machine